list_currencies better inside Invoices context
