+++ lib/siwapp_web/live/recurring_invoices_live/edit.ex
@@ -6,6 +6,7 @@ defmodule SiwappWeb.RecurringInvoicesLive.Edit do
   alias SiwappWeb.ItemView
 
   alias Siwapp.Commons
+  alias Siwapp.Invoices
   alias Siwapp.RecurringInvoices
   alias Siwapp.RecurringInvoices.RecurringInvoice
 
@@ -13,7 +14,7 @@ defmodule SiwappWeb.RecurringInvoicesLive.Edit do
     {:ok,
      socket
      |> assign(:series, Commons.list_series())
-     |> assign(:currency_options, SiwappWeb.PageView.list_currencies())
+     |> assign(:currency_options, Invoices.list_currencies())
      |> assign(:customer_suggestions, [])}
   end
