+++ lib/siwapp_web/views/page_view.ex
@@ -8,18 +8,4 @@ defmodule SiwappWeb.PageView do
     |> Money.new(currency)
     |> Money.to_string()
   end
-
-  def list_currencies do
-    default_currency = Siwapp.Settings.value(:currency)
-    Enum.uniq([default_currency] ++ primary_currencies() ++ all_currencies())
-  end
-
-  defp all_currencies do
-    Money.Currency.all()
-    |> Map.keys()
-    |> Enum.map(&Atom.to_string/1)
-    |> Enum.sort()
-  end
-
-  defp primary_currencies, do: ["USD", "EUR", "GBP"]
 end
